# Tweak NewsSite test step scenarios