--- a/resources/default-tests.mjs
+++ b/resources/default-tests.mjs
@@ -709,32 +709,44 @@ export const DefaultSuites = freezeSuites([
         },
         tests: [
             new BenchmarkTestStep("NavigateToUS", (page) => {
-                for (let i = 0; i < 25; i++) {
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                }
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-navlist-opinion-link").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
                 page.querySelector("#navbar-navlist-us-link").click();
                 page.layout();
             }),
             new BenchmarkTestStep("NavigateToWorld", (page) => {
-                for (let i = 0; i < 25; i++) {
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                }
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-navlist-opinion-link").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
                 page.querySelector("#navbar-navlist-world-link").click();
                 page.layout();
             }),
             new BenchmarkTestStep("NavigateToPolitics", (page) => {
-                for (let i = 0; i < 25; i++) {
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                }
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-navlist-opinion-link").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
                 page.querySelector("#navbar-navlist-politics-link").click();
                 page.layout();
             }),
@@ -749,32 +761,44 @@ export const DefaultSuites = freezeSuites([
         },
         tests: [
             new BenchmarkTestStep("NavigateToUS", (page) => {
-                for (let i = 0; i < 25; i++) {
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                }
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-navlist-opinion-link").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
                 page.querySelector("#navbar-navlist-us-link").click();
                 page.layout();
             }),
             new BenchmarkTestStep("NavigateToWorld", (page) => {
-                for (let i = 0; i < 25; i++) {
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                }
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-navlist-opinion-link").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
                 page.querySelector("#navbar-navlist-world-link").click();
                 page.layout();
             }),
             new BenchmarkTestStep("NavigateToPolitics", (page) => {
-                for (let i = 0; i < 25; i++) {
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                    page.querySelector("#navbar-dropdown-toggle").click();
-                    page.layout();
-                }
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-navlist-opinion-link").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
+                page.querySelector("#navbar-dropdown-toggle").click();
+                page.layout();
                 page.querySelector("#navbar-navlist-politics-link").click();
                 page.layout();
             }),