This proof of concept illustrates potential tweaks to the test steps for the NewsSite workload. It introduces an extra navigation to a new tab instead of repeatedly toggling the drop-down menu.

![image](https://github.com/user-attachments/assets/f83789ca-12da-476f-885e-13f47351eb08)