@camillobruni 2024-12-03
I'm not opposed to this change in general. @flashdesignory  what's your take here?
How much slower does the test run now?