@camillobruni 2024-12-03
btw, you can use [crossbench](https://chromium.googlesource.com/crossbench) to create browser comparison data easily.