@lingyuncai 2024-12-05
Hi @camillobruni, thanks for your comments.

Below are the crossbench results for this PR from 10 runs. We ran the whole benchmark and extracted the scores for NewsSite cases.

this PR
```
browser              Google Chrome        Firefox              Safari
NewsSite-Next        75.03 ± 0.92%        64.70 ± 0.90%        86.99 ± 0.61%
NewsSite-Nuxt        47.72 ± 0.98%        53.09 ± 1.2%         59.44 ± 0.92%
Score                36.45 ± 0.63%        31.74 ± 0.67%        31.41 ± 0.45%
```

main
```
browser              Google Chrome        Firefox              Safari
NewsSite-Next        64.35 ± 0.79%        63.03 ± 0.80%        71.6 ± 1.9%
NewsSite-Nuxt        62.70 ± 0.85%        70.6 ± 1.8%          76.33 ± 0.14%
Score                36.21 ± 0.53%        31.45 ± 0.90%        31.34 ± 0.59%
```

System and browser version
```
browser              Google Chrome        Firefox              Safari
version              131.0.6778.86        133.0                19618.1.15.11.12
os                   macos 14.4 arm64     macos 14.4 arm64     macos 14.4 arm64
device               Mac14,7              Mac14,7              Mac14,7
cpu                  Apple M2 8 cores     Apple M2 8 cores     Apple M2 8 cores
runs                 10                   10                   10
failed runs          0                    0                    0
```

Basically, NewsSite-Next runs slower while NewsSite-Nuxt runs faster. This difference is related to how each framework handles `router.push()`.

In NewsSite-Nuxt, the Vue router leverages microtasks for its navigation mechanism, so a forced layout of a new tab does not happen synchronously. However, in NewsSite-Next, `router.push()` is executed synchronously, resulting in a forced layout in sync time when clicking the drop-down menu again.