@lingyuncai 2024-12-18
Hi @camillobruni,

Just kindly ping :) Do you have any comments or suggestions on this idea? Thank you!